stego_api/requirementsAPI.txt
audioop-lts==0.2.1
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.2.1
cloudscraper==1.2.71
colorama==0.4.6
cryptography==44.0.2
decorator==5.2.1
Flask==3.1.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
m3u8==6.0.0
MarkupSafe==3.0.2
moviepy==2.2.1
numpy==2.3.2
pillow==11.3.0
proglog==0.1.12
pycparser==2.22
pycryptodome==3.23.0
pydub==0.25.1
pyOpenSSL==25.0.0
pyparsing==3.2.3
python-dotenv==1.1.1
requests==2.32.3
requests-toolbelt==1.0.0
six==1.17.0
tqdm==4.67.1
Unidecode==1.3.8
urllib3==2.3.0
Werkzeug==3.1.3
